[CH-3] Fixing an XSS vulnerability
We want to guard ourselves from malicious actors.

There are some problems with the implementation
around validating the integer input. I used the
Integer class to typecast and verify the input.
It raises an error when given bad arguments BUT
ALSO rounds floats. I should probably just verify
a match against a regex like /\d+/ ... but as they
say when you solve 1 problem with a regex ... you
end up with 2.

Other notable changes:
In our application helper I make a decent assumption
that our year inputs WILL ALWAYS be integers if they
are well formed requests.
I almost wrote a system test for asserting against
a bad request, but stopped since we already handle
that in the controller test.

Resolves [CH-3]